refactor: 도메인 서비스와 포트를 domain 계층으로 패키지 분리
Application Service(AuthService)와 Domain Service(AuthTokenService,
SocialLoginService)의 계층을 패키지로 명시한다.
- domain/service: 토큰 생명주기, 소셜 로그인 도메인 로직
- domain/port: OAuth 클라이언트 포트 인터페이스
- domain/model: AuthTokenResult 값 객체
- application/service: 유스케이스 조율 (AuthService)
